requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-openai
langchain-groq
langchain-ollama
pypdf
python-dotenv
sentence-transformers
python-multipart
faiss-cpu
reportlab
werkzeug
ruff
mypy